+++ scheme.js
@@ -5,11 +5,8 @@ import path from "path";
 
 const loadedTypes = loadFilesSync(path.join(__dirname, "./**/*.typeDefs.js"));
 const loadedResolvers = loadFilesSync(
-  path.join(__dirname, "./**/*.{queries,mutations}.js")
+  path.join(__dirname, "./**/*.resolvers.js")
 );
 
-const typeDefs = mergeTypeDefs(loadedTypes);
-const resolvers = mergeResolvers(loadedResolvers);
-
-const schema = makeExecutableSchema({ typeDefs, resolvers });
-export default schema;
+export const typeDefs = mergeTypeDefs(loadedTypes);
+export const resolvers = mergeResolvers(loadedResolvers);