+++ users/editProfile/editProfile.resolvers.js
@@ -1,0 +1,67 @@
+import fs from "fs";
+import client from "../../client";
+import bcrypt from "bcrypt";
+import { protectedResolver } from "../users.utils";
+import { GraphQLUpload } from "graphql-upload";
+
+export default {
+  Upload: GraphQLUpload,
+  Mutation: {
+    editProfile: protectedResolver(
+      async (
+        _,
+        {
+          firstName,
+          lastName,
+          username,
+          email,
+          password: newPassword,
+          bio,
+          avatar,
+        },
+        { loggedInUser }
+      ) => {
+        let hashedPassword = null;
+        if (newPassword) {
+          hashedPassword = await bcrypt.hash(newPassword, 10);
+        }
+        let avatarUrl = null;
+        if (avatar) {
+          const { filename, createReadStream } = await avatar;
+          const newFilename = `${loggedInUser.id}-${Date.now()}-${filename}`;
+          const readStream = createReadStream();
+          const writeStream = fs.createWriteStream(
+            process.cwd() + "/uploads/" + newFilename
+          );
+
+          readStream.pipe(writeStream);
+          avatarUrl = `http://localhost:4000/uploads/${newFilename}`;
+        }
+
+        const updatedUser = await client.user.update({
+          where: { id: loggedInUser.id },
+          data: {
+            firstName,
+            lastName,
+            username,
+            email,
+            bio,
+            ...(hashedPassword && { password: hashedPassword }),
+            ...(avatarUrl && { avatar: avatarUrl }),
+          },
+        });
+
+        if (updatedUser.id) {
+          return {
+            ok: true,
+          };
+        } else {
+          return {
+            ok: false,
+            error: "Could not update profile",
+          };
+        }
+      }
+    ),
+  },
+};