+++ users/login/login.resolvers.js
@@ -1,36 +1,9 @@
-import client from "../client";
+import client from "../../client";
 import bcrypt from "bcrypt";
 import jwt from "jsonwebtoken";
 
 export default {
   Mutation: {
-    createAccount: async (
-      _,
-      { username, firstName, email, password, lastName }
-    ) => {
-      try {
-        const existingUser = await client.user.findFirst({
-          where: {
-            OR: [{ username }, { email }],
-          },
-        });
-        if (existingUser) {
-          throw new Error("This username/email is already taken.");
-        }
-        const hashedPassword = await bcrypt.hash(password, 10);
-        return client.user.create({
-          data: {
-            firstName,
-            lastName,
-            username,
-            email,
-            password: hashedPassword,
-          },
-        });
-      } catch (e) {
-        return e;
-      }
-    },
     login: async (_, { username, password }) => {
       // find suer with args.username
       const user = await client.user.findFirst({